pyproject.toml
[project]
name = "summarize-rpg-session"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "openai>=1.68.2",
    "pyannote-audio>=3.3.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
